requirements.txt
Flask
gunicorn
sycopg2-binary
flask_sqlalchemy
flask_login
flask_wtf
python-dotenv
